removed/ renamed files. struvolpy for structure parsers
