--- powerfit.egg-info/PKG-INFO
@@ -1,14 +1,0 @@
-Metadata-Version: 2.1
-Name: powerfit
-Version: 1.1.0
-Summary: Rigid body fitting of high-resolution structures in low-resolution cryo-electron microscopy density maps
-Home-page: https://github.com/haddocking/powerfit
-Author: Gydo C.P. van Zundert
-Author-email: g.c.p.vanzundert@uu.nl
-License: Apache
-Classifier: Development Status :: 3 - Alpha
-Classifier: Programming Language :: Python :: 2.7
-Classifier: Intended Audience :: Science/Research
-Classifier: Topic :: Scientific/Engineering :: Bio-Informatics
-License-File: LICENSE
-License-File: NOTICE.txt
