--- powerfit.egg-info/SOURCES.txt
@@ -1,60 +1,0 @@
-.gitignore
-LICENSE
-NOTICE.txt
-README.md
-setup.py
-dist/powerfit-2.0.0-py3.7-linux-x86_64.egg
-powerfit/__init__.py
-powerfit/analyzer.py
-powerfit/elements.py
-powerfit/helpers.py
-powerfit/kernels.cl
-powerfit/powerfit.py
-powerfit/powerfitter.py
-powerfit/rotations.py
-powerfit/scripts.py
-powerfit/shape_descriptor.py
-powerfit/structure.py
-powerfit/volume.py
-powerfit.egg-info/PKG-INFO
-powerfit.egg-info/SOURCES.txt
-powerfit.egg-info/dependency_links.txt
-powerfit.egg-info/entry_points.txt
-powerfit.egg-info/requires.txt
-powerfit.egg-info/top_level.txt
-powerfit/data/E.npy
-powerfit/data/README
-powerfit/data/c48n309.npy
-powerfit/data/c48n527.npy
-powerfit/data/c48n9.npy
-powerfit/data/c48u1.npy
-powerfit/data/c48u1153.npy
-powerfit/data/c48u1201.npy
-powerfit/data/c48u1641.npy
-powerfit/data/c48u181.npy
-powerfit/data/c48u2219.npy
-powerfit/data/c48u27.npy
-powerfit/data/c48u2947.npy
-powerfit/data/c48u3733.npy
-powerfit/data/c48u4749.npy
-powerfit/data/c48u5879.npy
-powerfit/data/c48u7111.npy
-powerfit/data/c48u815.npy
-powerfit/data/c48u83.npy
-powerfit/data/c48u8649.npy
-powerfit/data/c600v.npy
-powerfit/data/c600vc.npy
-powerfit/tests/run_tests.py
-powerfit/tests/test__powerfit.py
-powerfit/tests/test_extensions.py
-powerfit/tests/test_gpyfft.py
-powerfit/tests/test_kernels.py
-powerfit/tests/test_powerfitter.py
-powerfit/tests/test_rotations.py
-powerfit/tests/test_shape_descriptor.py
-scripts/atom2dens.py
-scripts/fsc-curve.py
-scripts/rot_search.py
-src/_extensions.c
-src/_powerfit.c
-src/_powerfit.pyx
