--- powerfit.egg-info/entry_points.txt
@@ -1,4 +1,0 @@
-[console_scripts]
-em2em = powerfit.scripts:em2em
-image-pyramid = powerfit.scripts:image_pyramid
-powerfit = powerfit.powerfit:run
