--- powerfit.egg-info/requires.txt
@@ -1,2 +1,0 @@
-numpy>=1.8
-scipy
