--- powerfit.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-powerfit
